bits: Select for integers
